+++ block/src/formats/vhdx/internal/mod.rs
@@ -239,34 +239,16 @@ impl AsRawFd for Vhdx {
 #[cfg(test)]
 mod tests {
     use std::fs;
-    use std::process::Command;
-
-    use vmm_sys_util::tempfile::TempFile;
 
     use super::*;
-
-    /// Generate a small dynamic VHDX with `qemu-img`. Returns `None` (and the
-    /// test is skipped) when `qemu-img` is unavailable, e.g. in minimal CI.
-    fn dynamic_vhdx(size_mib: u64) -> Option<TempFile> {
-        let tf = TempFile::new().unwrap();
-        let path = tf.as_path();
-        let status = Command::new("qemu-img")
-            .args(["create", "-f", "vhdx", "-o", "subformat=dynamic"])
-            .arg(path)
-            .arg(format!("{size_mib}M"))
-            .status();
-        match status {
-            Ok(s) if s.success() => Some(tf),
-            _ => None,
-        }
-    }
+    use crate::formats::vhdx::test_util::create_dynamic_vhdx;
 
     /// An unaligned sector write under a forced O_DIRECT alignment must go
     /// through `AlignedFile`'s read-modify-write bounce (the data block and the
     /// BAT update both land at unaligned host offsets) and read back intact.
     #[test]
     fn unaligned_write_is_rmw() {
-        let Some(tf) = dynamic_vhdx(16) else {
+        let Some(tf) = create_dynamic_vhdx(16) else {
             eprintln!("skipping unaligned_write_is_rmw: qemu-img unavailable");
             return;
         };
@@ -299,7 +281,7 @@ mod tests {
 
     #[test]
     fn header_update_survives_reopen() {
-        let Some(tf) = dynamic_vhdx(16) else {
+        let Some(tf) = create_dynamic_vhdx(16) else {
             eprintln!("skipping header_update_survives_reopen: qemu-img unavailable");
             return;
         };
